New Leaderboard implementation
The new Leaderboard implementation creates a LeaderboardController and  Leaderboard class for memory persistence and keep a list of Players respectively. The old methods for saving scores were deleted from the Player class.
A new JDialog window is created for displaying the Leaderboard as Table.
The WindowView now displays the Highest Score in the Leaderboard as opposed to showing the Leaderboard.