Thread interruption implemented
The code in WindowView class no longer use the Thread.stop() method to interrupt threads. Inside ActionScreenView class a new boolean condition inside a while is checked instead to stop the current Thread. Now Threads can be stopped with the Thread.interrupt() method.

Also, a bug where the player could shoot after losing the game was fixed. The bug was also triggered when open a new window but reseting the focus to the ActionScreenView panel fixes it.