Rename and method to hide elements added
Changes in ActionScreenView class:

visual[] array is now visualElements[], size is by default 4 now.

Implementation of method setVisualElements() to set visual elements on or off.

Implementation of method getVisualElements() to get a boolean value of wether the visual elements are on or off.

Changes in WindowView class:

stop() method is now endCurrentGame()